feat(ui): Full help guide
